refactor: extract internal/logging
Phase A of #80 (split taskfileserver by concern). Moves the logging
primitives that have no Server coupling into a new internal/logging
package and rewires the call sites:

* internal/taskfileserver/logging.go -> internal/logging/fanout.go,
  exporting the previously-package-private fanoutHandler /
  newFanoutHandler as FanoutHandler / NewFanout. Adds Handlers() so
  external test code (the SDK arm assertion in logging_mcp_test.go)
  can walk the tree without poking unexported state.
* installMCPLogging on Server is replaced with logging.InstallMCP,
  invoked inline from HandleInitialized; LoggerName is now the
  package-level logging.LoggerName constant.
* parseLogLevel and newLogger from main.go become logging.ParseLevel
  and logging.NewLogger; main_test.go moves to
  internal/logging/level_test.go (git mv preserves history).
* TestFanoutHandler_DispatchesToAllHandlers and recordingHandler move
  with the source into internal/logging/fanout_test.go.

The Server-coupled tests (logging_test.go, the TestServer_* triplet
in logging_mcp_test.go) stay in internal/taskfileserver: they exercise
unexported helpers (Server.loadDesired, Server.log) that are not part
of the logging-package contract, so moving them would force either an
API surface increase or a test rewrite. They now reference
*logging.FanoutHandler in the type assertion.

No behaviour change: log levels, env var (MCP_TASKFILE_LOG_LEVEL),
structured field names, logger name, and handler ordering are
unchanged. task ci is green.

Refs #80
