+++ internal/logging/level_test.go
@@ -1,11 +1,11 @@
-package main
+package logging
 
 import (
 	"log/slog"
 	"testing"
 )
 
-func TestParseLogLevel(t *testing.T) {
+func TestParseLevel(t *testing.T) {
 	cases := []struct {
 		raw  string
 		want slog.Level
@@ -23,8 +23,8 @@ func TestParseLogLevel(t *testing.T) {
 	}
 	for _, tc := range cases {
 		t.Run(tc.raw, func(t *testing.T) {
-			if got := parseLogLevel(tc.raw); got != tc.want {
-				t.Errorf("parseLogLevel(%q) = %v, want %v", tc.raw, got, tc.want)
+			if got := ParseLevel(tc.raw); got != tc.want {
+				t.Errorf("ParseLevel(%q) = %v, want %v", tc.raw, got, tc.want)
 			}
 		})
 	}
